+++ cxl/acpi/chbs.c
@@ -3,7 +3,7 @@
 
 void display_chbs(struct chbs *chbs)
 {
-	printf("CHBS type %2d, len %4d, uid %4x, CXL version %2x, base %lx len %lx\n",
+	printf("CHBS type %d, len %d, uid %x, CXL version %x, base %lx len %lx\n",
 		chbs->type, chbs->record_length, chbs->uid,
 		chbs->cxl_version, chbs->base, chbs->length);
 }