+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to
 ## Unreleased
 #### Breaking Changes
 #### Added
-- libbpf: Add test `perf_buffer` symlink.
+- libbpf: Add test `ring_buffer`,`perf_buffer` symlink.
 #### Changed
 #### Deprecated
 #### Removed