+++ CHANGELOG.md
@@ -8,7 +8,21 @@ and this project adheres to
 
 
 ## Unreleased
+
 #### Breaking Changes
+#### Added
+#### Changed
+#### Deprecated
+#### Removed
+#### Fixed
+#### Security
+#### Docs
+#### Tools
+
+## [v2.2.13] 2025-06-19
+
+#### Breaking Changes
+- cxl: Support create vmem/pmem Qemu VM.
 #### Added
 - ntp: First commit.
 - oom: add cgroup-oom.sh.